Blank WordPress Theme
This is a blank theme I use for WordPress development. It has all of the
basic framework that I'll need. It also has more than just the two
widget locations that I find in many WP themes. In fact, it's got six!
It automatically adjusts the width of the content area based on whether
there's a left, right, or left and right sidebars. It's bootstrapped
(v2.3.2). It has no styling to it, which makes it ready to accept any
design that I am given. Finally, it's responsive, because pretty much
all themes these days need to be responsive.